+++ app/src/main/java/com/yzdsmart/Dingdingwen/bean/ScannedLog.java
@@ -10,14 +10,14 @@ public class ScannedLog {
     private String ImageUrl;//头像路径
     private String CSex;//用户性别
     private String CAge;//用户年龄
-    private Integer Gold;//获取金币
+    private Float Gold;//获取金币
     private String TimeStr;//时间间隔
     private String CreateTime;//扫码日期
 
     public ScannedLog() {
     }
 
-    public ScannedLog(String custCode, String CNickName, String imageUrl, String CSex, String CAge, Integer gold, String timeStr, String createTime) {
+    public ScannedLog(String custCode, String CNickName, String imageUrl, String CSex, String CAge, Float gold, String timeStr, String createTime) {
         CustCode = custCode;
         this.CNickName = CNickName;
         ImageUrl = imageUrl;
@@ -68,11 +68,11 @@ public class ScannedLog {
         this.CAge = CAge;
     }
 
-    public Integer getGold() {
+    public Float getGold() {
         return Gold;
     }
 
-    public void setGold(Integer gold) {
+    public void setGold(Float gold) {
         Gold = gold;
     }
